lib: Add x16/x32 optional trait functions for instruction interface
ARM9TDMI has a Harvard bus architecture with split instruction and data
interfaces (see DDI0168, "3.1 About the memory interface"). The
instruction interface memory accesses are exposed as "x16" and "x32"
functions in the Memory trait. Their implementation is optional, and
default to the current "r16" and "r32" functions.

This change might be useful when implementing cache/memory controllers
that discriminate code and data memory accesses.